Move  and  to
