+++ .gitattributes_specialized
@@ -1,5 +1,3 @@
-# Don't build the specialized modules in the standard release.
-specialized/ export-ignore
 tests/ export-ignore
 samples/ export-ignore
 utility/ export-ignore
