+++ ccc/specialized/array_adaptive_map.h
@@ -61,8 +61,8 @@ All types and functions can then be written without the `CCC_` prefix. */
 #include <stddef.h>
 /** @endcond */
 
+#include "../types.h" /* IWYU pragma: export */
 #include "private/private_array_adaptive_map.h"
-#include "types.h" /* IWYU pragma: export */
 
 /** @name Container Types
 Types available in the container interface. */
