+++ ccc/specialized/private/private_array_adaptive_map.h
@@ -20,7 +20,7 @@ limitations under the License.
 #include <stddef.h>
 /** @endcond */
 
-#include "../types.h"
+#include "../../types.h"
 
 /* NOLINTBEGIN(readability-identifier-naming) */
