+++ samples/words.c
@@ -27,9 +27,9 @@ Please specify a command as follows:
 #define TRAITS_USING_NAMESPACE_CCC
 #define TYPES_USING_NAMESPACE_CCC
 
-#include "ccc/array_adaptive_map.h"
 #include "ccc/flat_buffer.h"
 #include "ccc/sort.h"
+#include "ccc/specialized/array_adaptive_map.h"
 #include "ccc/traits.h"
 #include "ccc/types.h"
 #include "str_view/str_view.h"
