+++ tests/specialized/adaptive_map/test_adaptive_map_construct.c
@@ -3,11 +3,11 @@
 
 #define TRAITS_USING_NAMESPACE_CCC
 
-#include "adaptive_map.h"
 #include "adaptive_map_utility.h"
+#include "ccc/specialized/adaptive_map.h"
+#include "ccc/traits.h"
+#include "ccc/types.h"
 #include "checkers.h"
-#include "traits.h"
-#include "types.h"
 #include "utility/stack_allocator.h"
 
 static CCC_Adaptive_map
