+++ tests/specialized/adaptive_map/test_adaptive_map_iterator.c
@@ -10,8 +10,8 @@
 #define ADAPTIVE_MAP_USING_NAMESPACE_CCC
 
 #include "adaptive_map_utility.h"
-#include "ccc/adaptive_map.h"
 #include "ccc/flat_bitset.h"
+#include "ccc/specialized/adaptive_map.h"
 #include "ccc/traits.h"
 #include "ccc/types.h"
 #include "checkers.h"
