+++ tests/specialized/array_adaptive_map/test_array_adaptive_map_construct.c
@@ -5,7 +5,7 @@
 #define ARRAY_ADAPTIVE_MAP_USING_NAMESPACE_CCC
 
 #include "array_adaptive_map_utility.h"
-#include "ccc/array_adaptive_map.h"
+#include "ccc/specialized/array_adaptive_map.h"
 #include "ccc/traits.h"
 #include "ccc/types.h"
 #include "checkers.h"
