+++ tests/specialized/array_adaptive_map/test_array_adaptive_map_handle.c
@@ -7,11 +7,11 @@ the handle functions. */
 #define ARRAY_ADAPTIVE_MAP_USING_NAMESPACE_CCC
 #define TRAITS_USING_NAMESPACE_CCC
 
-#include "array_adaptive_map.h"
 #include "array_adaptive_map_utility.h"
+#include "ccc/specialized/array_adaptive_map.h"
+#include "ccc/traits.h"
+#include "ccc/types.h"
 #include "checkers.h"
-#include "traits.h"
-#include "types.h"
 
 static inline struct Val
 val(int const val) {
