+++ tests/specialized/array_adaptive_map/test_array_adaptive_map_iterator.c
@@ -10,8 +10,8 @@
 #define TYPES_USING_NAMESPACE_CCC
 
 #include "array_adaptive_map_utility.h"
-#include "ccc/array_adaptive_map.h"
 #include "ccc/flat_bitset.h"
+#include "ccc/specialized/array_adaptive_map.h"
 #include "ccc/traits.h"
 #include "ccc/types.h"
 #include "checkers.h"
